requirements.txt
Django==2.0
django-tinymce==2.7.0
pytz==2017.3
